pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "python-automation-services"
version = "0.0.1"

dependencies = [
    "pyyaml",
    "jsonschema",
    "easydict"
]

requires-python = ">=3.6"

